requirements.txt
# FastAPI - Modern, fast web framework for building APIs with Python
fastapi

# Uvicorn - ASGI server for running FastAPI applications
uvicorn[standard]

# Python-multipart - Required for handling file uploads in FastAPI
python-multipart

# PyPDF2 - Library for reading and manipulating PDF files
PyPDF2

# Pydantic - Data validation using Python type annotations
pydantic

# Python-dotenv - Library for loading environment variables from .env files
python-dotenv

# Pydantic-settings - For settings management
pydantic-settings

# NLTK - Natural Language Processing toolkit
nltk

# pdfplumber - Better PDF text extraction
pdfplumber

# python-docx - For Word document processing
python-docx

# MongoDB - Document database
motor

# Redis - Caching and session management
redis>=4.0.0

# AWS SDK - For S3 and other AWS services
boto3

# Additional dependencies for proper implementation
pymongo
httpx
python-jose[cryptography]
passlib[bcrypt]

# Render deployment
gunicorn 